refactor: Exception 위치 변경
